🎨 Ignore GPG output on signature verification
This plugin uses a dedicated keyring containing only the signing key
we know to trust because we obtained it from an official source via
a secure channel. However, because we haven't told GPG that we trust
the key, we see a trust warning when attempting to verify a signature
made using that key. This warning is spurious (aside: the GPG-level
trust model is esoteric and rarely used, in favor of pushing trust
upward to ensure you're only using trusted keys to verify signatures
in the first place).

If signature verification fails, we'll still get a non-zero exit code,
which will halt the process from continuing to install the errant file.